+++ cli/Tests/TuistLoaderTests/Loaders/PackageSettingsLoaderTests.swift
@@ -83,6 +83,7 @@ final class PackageSettingsLoaderTests: TuistUnitTestCase {
                 ],
                 defaultSettings: .recommended
             ),
+            expectedSignatures: [:],
             targetSettings: [:]
         )
         verify(manifestLoader)