+++ server/lib/tuist/tests/quarantined_test_case.ex
@@ -12,6 +12,7 @@ defmodule Tuist.Tests.QuarantinedTestCase do
     :quarantined_by_account_id,
     :quarantined_by_account_name,
     :last_ran_at,
-    :last_run_id
+    :last_run_id,
+    :last_status
   ]
 end